Not using dato-react-ui dropdown, rolling our own based on actual html select to avoid padding outside plugin.
