+++ src/components/Drowdown/styles.module.css
@@ -1,0 +1,34 @@
+.select {
+  width: 100%;
+  border: 1px solid var(--border-color);
+  -moz-appearance: none;
+  -webkit-appearance: none;
+  outline: none;
+  appearance: none;
+  font-size: 0.9375rem;
+  color: rgb(52, 54, 58);
+  padding: 10px;
+  background-color: #fff;
+  border-radius: 0;
+  background-image: url('/public/assets/chevron.svg');
+  background-size: 10px 10px;
+  background-repeat: no-repeat;
+  background-position: right var(--spacing-m) center;
+  transition: 0.2s cubic-bezier(0.55, 0, 0.1, 1);
+}
+
+svg {
+  color: #000;
+}
+
+svg path {
+  color: #000;
+}
+
+select:hover {
+  border-color: #d7d7d7;
+}
+
+select::-ms-expand {
+  display: none; /* Hide the default arrow in Internet Explorer 10 and Internet Explorer 11 */
+}
