--- src/entrypoints/LocaleSelector/styles.module.css
@@ -1,3 +1,0 @@
-.form {
-  min-height: 100px;
-}
